Implemented Hypercube Search

+++ lib/utils/inc/mathUtils.h
@@ -13,7 +13,7 @@ double euclideanDistance(PointPtr x, PointPtr y, int dimension);
 double min(double x1, double x2, double x3);
 int minIndx(double x1, double x2, double x3);
 double DFDistance(const PointPtr p, const PointPtr q, int dimension, std::vector<std::vector<double>> *_c = NULL);
-int avoidOverFlowModulo(long a, long b, long m, char op);
+long avoidOverFlowModulo(long a, long b, long m, char op);
 long euclideanModulo(long x, long y);
 int powerWithBase2(int exp);
 bool is_number(const std::string &s);
